Add ability to insert rows and columns
