+++ include/internal/data_frame/fwd.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace csv {
+    template<typename KeyType>
+    class DataFrame;
+    template<typename KeyType>
+    class DataFrameColumn;
+    template<typename KeyType>
+    class DataFrameRow;
+}
