+++ include/internal/memory/raw_csv_field.hpp
@@ -5,10 +5,8 @@
 #pragma once
 
 #include <cassert>
-#include <cstdint>
 
 #include "../common.hpp"
-#include "constants.hpp"
 
 namespace csv {
     namespace internals {
@@ -21,18 +19,18 @@ namespace csv {
                     size_t _length,
                     bool _is_realized = false
                 ) noexcept {
-                    assert(_start <= INVALID_REALIZED_OFFSET);
-                    assert(_length <= INVALID_REALIZED_OFFSET);
-                    start = static_cast<std::uint32_t>(_start);
-                    length = static_cast<std::uint32_t>(_length);
+                    assert(_start <= CSV_CHUNK_INDEX_MAX);
+                    assert(_length <= CSV_CHUNK_INDEX_MAX);
+                    start = static_cast<CSVChunkIndex>(_start);
+                    length = static_cast<CSVChunkIndex>(_length);
                     is_realized = _is_realized;
                 }
 
                 /** Raw row-relative start, or quote-arena logical start when is_realized is true. */
-                std::uint32_t start = 0;
+                CSVChunkIndex start = 0;
 
                 /** Field length in the selected backing storage. */
-                std::uint32_t length = 0;
+                CSVChunkIndex length = 0;
 
                 /** True when start/length refer to RawCSVData::quote_arena instead of RawCSVData::data. */
                 bool is_realized = false;
